+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * Items that differ only in their components have been added to groups
+* The order of items has been changed to the default one